Backend

+++ Code/backend/app/models/intelligence.py
@@ -3,10 +3,10 @@ from pydantic import BaseModel
 
 class IntelligencePayload(BaseModel):
     vehicle_id: str
-    timestamp_ms: int
+    timestamp_ms: float   # allows int or float
 
-    fog_decision_critical_class: int
-    fog_decision_actuation_triggered: int
+    fog_decision_critical_class: float
+    fog_decision_actuation_triggered: float
     fog_decision_confidence: float
 
     thermal_brake_margin: float
@@ -14,15 +14,15 @@ class IntelligencePayload(BaseModel):
     thermal_stress_index: float
 
     mechanical_vibration_anomaly_score: float
-    mechanical_dominant_fault_band_hz: int
+    mechanical_dominant_fault_band_hz: float
     mechanical_vibration_rms: float
 
     electrical_charging_efficiency_score: float
-    electrical_battery_health_pct: int
+    electrical_battery_health_pct: float
 
-    engine_rul_pct: int
-    brake_rul_pct: int
-    battery_rul_pct: int
+    engine_rul_pct: float
+    brake_rul_pct: float
+    battery_rul_pct: float
 
     vehicle_health_score: float
